Week1 and Week2 (#2)
* Create CourseDevelopmentWeek1.MD

Describtion of first week of the course.

* Rename CourseDevelopmentWeek1.MD to CourseDevelopment/Week1.MD

* Rename CourseDevelopment/Week1.MD to CourseDevelopment/Week1/Description.MD

* Create SyllabusWeek2.MD

* Rename Description.MD to Chapter1.MD

* Rename SyllabusWeek2.MD to Chapter2.MD

* Add files via upload

* Update Chapter1.MD

* Add files via upload

* Update Chapter1.MD

* Add files via upload

* Update Chapter1.MD

Added pictures and fine tuning

* Add files via upload

This is a rolling slide deck., ones in a while it will be updated here. It is used to support the creation process of the course development. The week specific .MD-files are a summary of some slides in the deck.

* Update Chapter1.MD

* Add files via upload

Frontpage of Chapter 2

* Add files via upload

* Update Chapter2.MD

Added week2 Lecture and Tutorial ideas

* Add files via upload