Create Readme.MD
Hi,

Hereby my ideas for week 8 of the course. In my opinion it is wise to start with a short lecture. After that, we can organize a 'Future (Scientific) Job Information Market'. The grades will be based on mind-maps the students will make. If time is with us, it is possible to let students do an elavator pitch. Hope you like it!

Best,

Hendrik
'