Create Week6.MD (#8)
* Create Week6.MD

Hi Sanli and Erik,

Here are my ideas for week 6 of the course based on the keywords mentioned in the syllabus and conversations I have had with colleagues at the UU. The lecture I consider as theoretical, while the tutorial is practical oriented. Also links with other weeks, such as Big Science (week 3), can be made. Note that this week, especially the tutorial will need a lot of preparation. My advice is therefore to do a test run with a few students. 

Best,

Hendrik

* Rename Week6.MD to Readme.MD