Create Readme.MD (#10)
* Create Readme.MD

Hi,

Hereby my ideas for week 8 of the course. In my opinion it is wise to start with a short lecture. After that, we can organize a 'Future (Scientific) Job Information Market'. The grades will be based on mind-maps the students will make. If time is with us, it is possible to let students do an elavator pitch. Hope you like it!

Best,

Hendrik
'

* Rename CourseDevelopment/Readme.MD to CourseDevelopment/Week8/Readme.MD

Changed Folder

* Update Readme.MD

Added e-science centre

* Update Readme.MD

- Goal of the week changed (the goal of week6 was there by accident) 
- line 61, the word traditional is avoided. 
- Footnote on PhD Advantage

* Update Readme.MD

A new option in the tutorial session is created, for the case where the visiting professionals do not have much time.