Create Readme.MD
Hi,

In this proposal Week 7 consists out of a lecture and a tutorial. In the tutorial groups of students will create a new, or study an existing, open businesses. In the end every student hands in a canvas with personal notes and  an analysis of the open business. This will be graded with a pass/fail. 

Best,

Hendrik