Create Readme.MD (#30)
* Create Readme.MD

Hi,

A first draft of week 4. I am not so familiar with the topics of this week, hence it is strongly recommended, to let an experienced person further develop the described assignments in detail. 
Although, it is not my expertise I think that these topics may be relevant for future physicists. It is a bit dry course material, so a large focus on the tutorial session can help students the most. To let them prepare for future research projects it may be relevant to: create a folder structure of an arbitrary open science project. The same holds for setting up a systematic review or a python environment. 

Best,

Hendrik

* Update Readme.MD

Added a new assignment, rewritten the goal of the assignment and structured the assignment into 3 parts.

* Update Readme.MD

added an article of Cremari on the use of colours in scientific communication:
 Crameri, F., Shephard, G.E. & Heron, P.J. The misuse of colour in science communication. Nat Commun 11, 5444 (2020). https://doi.org/10.1038/s41467-020-19160-7